requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.2
click==8.0.3
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crcmod==1.7
cryptography==3.4.8
dashscope==1.23.4
dataclasses-json==0.6.7
datasets==2.18.0
dbus-python==1.2.18
dill==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.3.0
filelock==3.18.0
frozenlist==1.6.2
fsspec==2024.2.0
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.10.0
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-modelscope-integration==0.1.0
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langsmith==0.3.44
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
marshmallow==3.26.1
modelscope==1.26.0
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.14
mypy_extensions==1.1.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
openai==1.84.0
orjson==3.10.18
oss2==2.19.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pyarrow==20.0.0
pyarrow-hotfix==0.7
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycryptodome==3.23.0
pycurl==7.44.1
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2022.1
PyYAML==5.4.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.18.0
safetensors==0.5.3
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
ssh-import-id==5.11
sympy==1.14.0
systemd-python==234
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
triton==3.3.0
Twisted==22.1.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
websocket-client==1.8.0
xxhash==3.5.0
yarl==1.20.0
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.23.0
